Fix coverity issues, 497452, 497428, 497414, 497443, 497438
497452, 497428, 497414, Double free.
497443, An integer overflow.
497438, An integer overflow.

+++ crypto/asn1/x_delegated_credential.c
@@ -202,7 +202,7 @@ DELEGATED_CREDENTIAL *DC_load_from_file_ex(const char *file,
     BIO *bio_dc = NULL;
     char *dc_hex_buf = NULL;
     unsigned char *dc_buf = NULL;
-    size_t dc_hex_len, len;
+    int dc_hex_len, len;
     size_t dc_buf_len;
 
     dc_hex_buf = OPENSSL_malloc(DC_MAX_LEN);
@@ -233,7 +233,7 @@ DELEGATED_CREDENTIAL *DC_load_from_file_ex(const char *file,
      */
     len = dc_hex_len / 2;
 
-    dc_buf = OPENSSL_malloc(len);
+    dc_buf = OPENSSL_malloc((size_t)len);
     if (dc_buf == NULL) {
         ERR_raise(ERR_LIB_ASN1, ERR_R_MALLOC_FAILURE);
         goto err;
